Adding ability to replace all send buttons
This is a very crude implimentation at this point.  It will find every
send button (including the new compose windows that are generated on the
fly) and replace the send button with a custom one.  Right now it just
displays an alert and then still submits the form, but this will be
updated...